+++ CHANGES.rst
@@ -9,3 +9,8 @@ Changes
 
 - The matchers in ``nti.schema.testing`` have been moved to
   ``nti.testing.matchers``.
+- Using ``AdaptingFieldProperty`` will now raise the more specific
+  ``SchemaNotProvided`` error instead of a ``TypeError`` if adapting
+  the value fails.
+- ``EqHash`` has moved from ``nti.schema.schema`` to
+  ``nti.schema.eqhash``. A compatibility shim remains.